pyproject.toml
[project]
name = "siglip-from-scratch"
version = "0.1.0"
description = "PyTorch implementation of SigLIP (inference only)" 
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dataclasses>=0.8",
    "matplotlib>=3.10.3",
    "pillow>=11.2.1",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "transformers>=4.51.3",
]
